requirements.txt
pygame~=2.6.1
PyOpenGL~=3.1.7
python-dotenv~=1.0.1